--- src/constants/ticks.ts
@@ -1,9 +1,0 @@
-import JSBI from 'jsbi'
-
-// Tick with fields parsed to JSBIs, and active liquidity computed.
-export interface TickProcessed {
-  tickIdx: number
-  liquidityActive: JSBI
-  liquidityNet: JSBI
-  price0: string
-}
