+++ src/utils/computeSurroundingTicks.ts
@@ -1,7 +1,7 @@
 import { Token } from '@uniswap/sdk-core'
 import { tickToPrice } from '@uniswap/v3-sdk'
+import { TickProcessed } from 'hooks/usePoolTickData'
 import JSBI from 'jsbi'
-import { TickProcessed } from 'constants/ticks'
 import { AllV3TicksQuery } from 'state/data/generated'
 
 const PRICE_FIXED_DIGITS = 8
